+++ src/OrchardCore/OrchardCore.Search.AzureAI.Core/Deployment/AzureAISearchIndexDeploymentSource.cs
@@ -2,6 +2,7 @@ using System.Linq;
 using System.Text.Json.Nodes;
 using System.Threading.Tasks;
 using OrchardCore.Deployment;
+using OrchardCore.Search.AzureAI.Deployment.Models;
 using OrchardCore.Search.AzureAI.Models;
 using OrchardCore.Search.AzureAI.Services;
 
@@ -54,19 +55,4 @@ public class AzureAISearchIndexDeploymentSource(AzureAISearchIndexSettingsServic
             IndexLatest = settings.IndexLatest,
             Culture = settings.Culture,
         };
-}
-
-public class AzureAISearchIndexInfo
-{
-    public string IndexName { get; set; }
-
-    public string AnalyzerName { get; set; }
-
-    public string QueryAnalyzerName { get; set; }
-
-    public bool IndexLatest { get; set; }
-
-    public string[] IndexedContentTypes { get; set; }
-
-    public string Culture { get; set; }
 }
